+++ acoustics/FemModule.h
@@ -1,0 +1,87 @@
+﻿// -*- tab-width: 2; indent-tabs-mode: nil; coding: utf-8-with-signature -*-
+//-----------------------------------------------------------------------------
+// Copyright 2000-2024 CEA (www.cea.fr) IFPEN (www.ifpenergiesnouvelles.com)
+// See the top-level COPYRIGHT file for details.
+// SPDX-License-Identifier: Apache-2.0
+//-----------------------------------------------------------------------------
+/*---------------------------------------------------------------------------*/
+/* FemModule.h                                                 (C) 2022-2024 */
+/*                                                                           */
+/* Simple module to test simple FEM mechanism.                               */
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+#ifndef FEMMODULES_H
+#define FEMMODULES_H
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+#include <arcane/utils/NumArray.h>
+#include <arcane/utils/CommandLineArguments.h>
+#include <arcane/utils/StringList.h>
+
+#include <arcane/ITimeLoopMng.h>
+#include <arcane/IMesh.h>
+#include <arcane/IItemFamily.h>
+#include <arcane/ItemGroup.h>
+#include <arcane/ICaseMng.h>
+
+#include "IDoFLinearSystemFactory.h"
+#include "Fem_axl.h"
+#include "FemUtils.h"
+#include "DoFLinearSystem.h"
+#include "FemDoFsOnNodes.h"
+#include "ArcaneFemFunctions.h"
+
+/*---------------------------------------------------------------------------*/
+/*---------------------------------------------------------------------------*/
+
+using namespace Arcane;
+using namespace Arcane::FemUtils;
+
+/*---------------------------------------------------------------------------*/
+/**
+ * @brief A module for finite element method.
+ *
+ * This class handles the initialization and computation for finite element
+ * method (FEM) simulations, providing methods to  set  up and solve linear
+ * systems, assemble FEM operators, and perform result checks.
+ */
+/*---------------------------------------------------------------------------*/
+
+class FemModule
+: public ArcaneFemObject
+{
+ public:
+
+  explicit FemModule(const ModuleBuildInfo& mbi)
+  : ArcaneFemObject(mbi)
+  , m_dofs_on_nodes(mbi.subDomain()->traceMng())
+  {
+    ICaseMng* cm = mbi.subDomain()->caseMng();
+    cm->setTreatWarningAsError(true);
+    cm->setAllowUnkownRootElelement(false);
+  }
+
+  void startInit() override; //! Method called at the beginning of the simulation
+  void compute() override; //! Method called at each iteration
+  VersionInfo versionInfo() const override { return VersionInfo(1, 0, 0); }
+
+ private:
+
+  Real m_kc2;
+
+  DoFLinearSystem m_linear_system;
+  IItemFamily* m_dof_family = nullptr;
+  FemDoFsOnNodes m_dofs_on_nodes;
+
+  void _doStationarySolve();
+  void _getMaterialParameters();
+  void _assembleBilinearOperatorTRIA3();
+  void _solve();
+  void _assembleLinearOperator();
+  void _validateResults();
+
+  FixedMatrix<3, 3> _computeElementMatrixTRIA3(Cell cell);
+};
+
+#endif
